Tercer commit

+++ src/TiendaMVC/base/Producto.java
@@ -59,7 +59,7 @@ public class Producto {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
